requirements.txt
astroid==3.0.3
blinker==1.7.0
click==8.1.7
colorama==0.4.6
dill==0.3.8
flake8==7.0.0
Flask==3.0.2
isort==5.13.2
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.5
mccabe==0.7.0
pillow==10.2.0
platformdirs==4.2.0
pycodestyle==2.11.1
pyflakes==3.2.0
python-barcode==0.15.1
tomlkit==0.12.3
Werkzeug==3.0.1
